Generate separate integration module and use AS7's automatic activation
The build now generates two modules:

  org.eclipse.persistence
  id.au.ringerc.as7.eclipselinkintegration

The former is simply EclipseLink with no funny business except for adding
an optional dependency on the latter.

The latter module adds a services loader specifying that it should be
activated as a persistence provider adapter. It will be auto-loaded
by JBoss AS 7 if it is present, and will add appropriate properties
to integrate its self into EclipseLink transparently to the user.